Moved as much initialization out of static scope as possible
Static initialization was causing a lot of trouble with the STMduino library. Plus, it makes the code unpredictable so it was best to move initialization out of the static scope.

Added init() virtual method to Singleton
Init all Singletons at the beginning of the program.

Moved pin initialization out of ComMgr to the beginning of the program to be clearer.

Added serial outputs to initialization.

Changed HardwareTimer for steppers as TIM7 appears to be taken by the Dynamixel SoftwareSerial. (TO CHECK FOR PWMs LATER)